reorganize repo for more modularity
